Set max-width and center main content
Added a new class "content" to adjust the main content visual on the website. This class sets the max-width to 1200px and centers this content using a margin 0 auto. The "content" class was inserted in several html pages ensuring the changes are applied across the website. This commit enhances the layout adjustment and consistency of the main content display.