Reorder table columns and refactor CSS styling in basket page
The table columns have been reordered in the basket product list – the amount is now displayed after the item name. Additionally, the CSS for the table has been refactored, changing widths and alignments of table header and cells along with the implementation of media queries for responsive design. Changes aim to improve readability and adaptability of the basket page's table.